RFC: arm: Create boot image header
In order to support a bootloader, with possibly signed images, we need
to be able to determine how large the image itself is.  The Cortex M
CPUs generally require the vector table to be at the beginning of the
address space.  Put this image header immediately after the vector
table.

This header contains some magic numbers (generated randomly) followed by
some of the linker symbols marking the beginning and end of the sections
in flash.  With some basic arithmetic on these values, the bootloader
can determine how large the flash (and signed part of) image is.

Change-Id: I73b95c5226ee4a959b4f65d6b8b949cd52f9fca0
Signed-off-by: David Brown <david.brown@linaro.org>