WIP: Working RSA signature verification
Verify an appended RSA signature.  The README describes the steps to
producing the key, and embedding the root key into the bootloader.

The sign.sh script will append the signature to the shell image (done in
place).

There are a few hacks to the mbedtls configuration to enable RSA and
increase the numeric sizes so that this all works.

Additionally, we implement is very naive 'malloc/free' implementation,
that just allocates from a buffer (and never frees).  Given a large
enough buffer, this is adequate to perform the RSA signature
verification.

Some notes.  Mbedtls is built by default for EC signatures only.  This
is worth investigating, as the working values are significantly smaller
(typically about 1/8 the size).  The mynewt bootloader has code to do an
EC verification, but they don't describe how to append this kind of
signature to the image.

Although this malloc workaround is adequate to make this run, it is
unclear if this is desired.  To verify multiple signatures, we could
just reset the buffer at the end, and redo the verification.  Since the
verification is likely to be the only use of dynamic allocation, this
will probably work.  Another possibility would be to investigate one of
the allocators within Zephyr (there is a general heap allocator).