WIP: More forward thinking signature format
Instead of just a bunch of bytes at the end of the image, the signature
will itself have a header.  Included in this header is a magic number,
and some fields indicating the algorithm and format of the signature.
There is also a length, because some signatures can vary in length.