boot: Add support for RSA signatures
Images can be signed with either ECDSA or RSA.  Add support for the RSA
signatures.  Note that as defined now, the mbedTLS library won't compile
if both are enabled.

This change also checks in a sample RSA key (root.pem) that can be used
for development.  Obviously, this key shouldn't be used in production,
as the private key is publically available.

Change-Id: Ife1e453f1f64733603de8c136494d40a9f93ad1f
Signed-off-by: David Brown <david.brown@linaro.org>