boot: Simple flash image update
Detect a signed update image in a configurable upgrade partition, and if
present, replace the primary image with this new image.  Because this
procedure is fairly simple (if upgrade is present, always use it, and
don't erase the upgrade until the new image has been flashed), this
upgrade process should be fairly robust.  As long as the new image has
been tested, this should work reliably.

Change-Id: I7b1836202972212c9876babd530633132bd3b459
Signed-off-by: David Brown <david.brown@linaro.org>