Create CONFIG_BOOT_HEADER
The Mynewt project bootloader prepends a small header to the image, and
expects the image to be built and liked such that it can run with this
header in place.  Since Zephyr leaves the boot vector table in flash, we
need to know about this here.

Enabling CONFIG_BOOT_HEADER will prepend 128 bytes of zeros to the
image, moving the vector table to the next possible slot.  It is
expected for external tools to be able to replace this padding with the
header itself.

This also fixes the stm32f4 target, which assumes the vector table
resides at the start of flash.