arm: mps_an521: Add support for the AN521 SoC with MPS2 board
The AN521 is a Soft Macro Model implementation of the SSE-200 subsystem
with SIE-200 and CMSDK components targeting the MPS2+ board.  The
SSE-200 subsystem implements two Cortex-M33 cores.

ARM's Cortex-M Prototyping System (MPS2) is a board with an FPGA that
can be programmed with different 'SoCs'. To use these in Zephyr we
need a set of board files for each variant.

This adds a board for a variant which implements a Cortex-M33 CPU; the
naming of this matches that used for the Zephyr soc (which is itself
based on ARM's documentation nomenclature).

Signed-off-by: Kumar Gala <kumar.gala@linaro.org>
Signed-off-by: Karl Zhang <karl.zhang@linaro.org>