lib: jwt: Add jwt library
Currently, this is in drivers, because things in lib don't seem to be
able to find the mbedTLS includes.  So, for this demo, let's call it a
'driver' since that does seem to be able to get to the TLS code.

There is a unit test, but it just generates a token, and it is up to the
user to verify that it works.  There is also quite a bit of stack/ram
usage that perhaps could be reduced.