samples: rust: Start of a small rust sample
This is a minimalist rust hello-world program for Zephyr.  Currently, many
things are just hard-coded in the CMakeLists.txt file, including the target
directory, and the build target.  However, this is able to have cmake
invoke `cargo build` to build the rust code into a library, which is then
linked into the program.

Signed-off-by: David Brown <david.brown@linaro.org>